requirements.txt
aiosmtplib==1.1.6
anyio==3.3.4
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.8
django-appconf==1.0.5
django-phonenumber-field==5.2.0
django-user-accounts==3.0.4
h11==0.12.0
httpcore==0.13.7
httpx==0.19.0
idna==3.3
messages==0.8.0
phonenumbers==8.12.36
pytz==2021.3
rfc3986==1.5.0
sniffio==1.2.0
sqlparse==0.4.2
validus==0.3.0
